+++ src/scenes/posts/index.jsx
@@ -1,4 +1,5 @@
-import { Box, Button, useTheme } from "@mui/material";
+/* eslint-disable no-restricted-globals */
+import { Box,Modal, Button, useTheme } from "@mui/material";
 import { DataGrid, GridToolbar } from "@mui/x-data-grid";
 import { tokens } from "../../theme";
 import CheckBoxRoundedIcon from "@mui/icons-material/CheckBoxRounded";
@@ -6,16 +7,23 @@ import DangerousOutlinedIcon from "@mui/icons-material/DangerousOutlined";
 import AddOutlinedIcon from "@mui/icons-material/AddOutlined";
 import Header from "../../components/Header";
 import { useEffect } from "react";
-import { fetchPosts, delete_post } from "../../config/services/api_calls";
-import { useState } from "react";
+import { fetchPosts, delete_post, fetchMyPosts } from "../../config/services/postServices";
+import { useState,useContext } from "react";
 import { useNavigate } from "react-router-dom";
 import SimpleSnackbar from "../global/snackbar";
+import AuthContext from "../../config/store/auth-context";
+import { baseUrl } from "../../config/api/apiHelpers";
+
 const Posts = (props) => {
+  const authCtx = useContext(AuthContext);
   // variable definations
   const navigate = useNavigate();
   const [posts, setPosts] = useState([]);
+  const [selectedPost, setSelectedPost] = useState([]);
   const theme = useTheme();
   const colors = tokens(theme.palette.mode);
+  const [open, setOpen] = useState(false);
+
   const [snak, setsnak] = useState({
     severity: "",
     message: "",
@@ -23,14 +31,25 @@ const Posts = (props) => {
   });
   // functions
   useEffect(() => {
-    fetchPosts().then((res) => {
-      if (res.success && res.data) {
-        setPosts(res.data.posts);
-      } else {
-        console.log(res.error);
-      }
-    });
-  }, []);
+    if(authCtx.role === "BUSSINES_ACCOUNT"){
+      fetchMyPosts().then((res) => {
+        if (res.success && res.data) {
+          setPosts(res.data);
+        } else {
+          console.log(res.error);
+        }
+      });
+    }else if(authCtx.role === "ADMIN" || authCtx.role ===  "SUPER_ADMIN"){
+      fetchPosts().then((res) => {
+        if (res.success && res.data) {
+          setPosts(res.data.posts);
+        } else {
+          console.log(res.error);
+        }
+      });
+    }
+
+  }, [authCtx.role]);
   const handleClose = () => {
     setsnak({
       open: false,
@@ -41,35 +60,82 @@ const Posts = (props) => {
   const editHandler = (u_id) => {
     navigate(`/editpost/${u_id}`);
   };
+
   const deleteHandler = (postID) => {
     props.isloading(10);
-    delete_post(postID).then((res) => {
-      console.log(res.data);
-      if (res.success && res.data) {
-        setsnak({ severity: "success", message: res.data.message, open: true });
-      } else {
-        setsnak({
-          severity: "error",
-          message: res.error,
-          open: true,
-        });
-        console.log(res.error);
-      }
-    });
+    if(confirm("Are you sure you want to delete this post?") === true){
+      delete_post(postID).then((res) => {
+        if (res.success && res.data) {
+          setsnak({ severity: "success", message: res.data.message, open: true });
+        } else {
+          setsnak({
+            severity: "error",
+            message: res.error,
+            open: true,
+          });
+          console.log(res.error);
+        }
+      });
+    }
     props.isloading(100);
   };
 
+
+  const handleModalOpen = (post) => {
+    setSelectedPost(post);
+    setOpen(true)
+  };
+  const handleModalClose = () => setOpen(false);
+
   const columns = [
+    {
+      field: "thumbnail",
+      headerName: "Thumbnail",
+      flex: 1,
+      renderCell:(params)=>{
+        return (
+          <Box>
+            <Button
+              onClick={() => handleModalOpen(params.row)}
+              color="secondary"
+              variant="outlined"
+            >
+              Thumbnail
+            </Button>
+            <Modal
+              keepMounted
+              open={open}
+              onClose={handleModalClose}
+              aria-labelledby="User Detail Profile"
+            >
+              <Box sx={{
+                position: 'absolute',
+                top: '50%',
+                left: '50%',
+                transform: 'translate(-50%, -50%)',
+                width: 600,
+                bgcolor: colors.blueAccent[900],
+                boxShadow: 24,
+                p: 4,
+              }}>
+                <img src={`${baseUrl}/${selectedPost.thumbnail}`} alt="" width="500"  />
+              </Box>
+            </Modal>
+          </Box>
+          
+        )
+      }
+    },
     {
       field: "title",
       headerName: "Title",
       cellClassName: "name-column--cell",
-      flex: 2,
+      flex: 1,
     },
     { field: "description", headerName: "Description", flex: 2 },
     {
       field: "is_approved",
-      headerName: "Approved",
+      headerName: "Published",
       renderCell: (params) => {
         return (
           <Box display="flex" justifyContent="center" alignItems="center">
@@ -98,7 +164,7 @@ const Posts = (props) => {
               color="secondary"
               variant="outlined"
             >
-              Update
+              Detail
             </Button>
             <Button
               onClick={() => deleteHandler(params.row._id)}
